Renamed the variables/functions
The variables and functions now fit the athlete or not purpose of the code. I also cleaned up some of the other code so it fit the new purpose like changing the questions and the variables throughout the code. I also tried to add the UT logo to the first page.